Build a Libvirt ppc64 container image
Using the previously added RPMs, build a ppc64 container image.
Files were changed without too much regard to other architectures.

Dockerfile.in was changed to disable copr (it's not enabled
for ppc64). The RPMs are copied to the container to be installed
directly.

Steps to initialize the docker daemon got commented out in
prepare.sh.

'--network host' was added in a few places due to network errors
I was experiencing with the container not being able to connect
to external sources. This is also the case with the buildkitd-flags
added in 'docker buildx create' command line in prepare.sh. I
ended up changing the local firewall via firewall-cmd later on, so
perhaps this is not needed.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>